Fix request path and query params' example support.
The adoc output of a path' example always use the default value.For
example in the unit test:

* pets/{petId} use the path params `petId` with example 30, but the
output adoc rendered pets/0 (0 is the default value of int param)
* pets/{findByTags} use the query params `tags` with example adorable,
but the output adoc rendered "tags":"string" ("string" is the default
value of string param)

It should use the "x-example" of the swagger(also the method `example()`
of the swagger-core jar.
This is the first commit to fix this problem, which add the
unit test for it.